fixed Draft implementation of LLM assistant // TODO: Database construction for user history input. Abstract for conversations. Draggable documents(RAG mode/Context mode). Search Engine(with Analysis agent). Report exporter (PDF/md/txt/xml) -> workspace for human validator? Highlighter (for code).
